Schedule the autosave from the field change handlers like the other settings forms
